requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Optional: Transformers (for BERT)
# Uncomment to use BERT classifier
# transformers>=4.30.0
# torch>=2.0.0

# Optional: LLM providers
# Uncomment to use specific LLM providers
# openai>=1.0.0
# anthropic>=0.5.0

# Optional: API server
# Uncomment to run API server
# fastapi>=0.100.0
# uvicorn>=0.23.0
# pydantic>=2.0.0

# Development
pytest>=7.0.0
